Changed Rule Validations in UserRequest

+++ app/Http/Controllers/Api/UserController.php
@@ -108,18 +108,31 @@ class UserController extends Controller
     {
         $validated = $request->validated();
 
-        // Upload Image to Backend and Store Image Path
-        $validated['image'] = $request->file('image')->storePublicly('user', 'public');
+        // Check if a file was uploaded
+        if ($request->hasFile('image')) {
 
+            // Upload Image to Backend and Store Image Path
+            $validated['image'] = $request->file('image')->storePublicly('user', 'public');
+
+            // Get Info by Id 
+            $user = User::findOrFail($id);
+
+            // Delete Previous Image
+            if (!is_null($user->image)) {
+                Storage::disk('public')->delete($user->image);
+            }
+
+            // Update New Info
+            $user->update($validated);
+
+            return $user;
+        }
+
+        // If no file was uploaded or if the file is not valid, proceed without updating the image
         // Get Info by Id 
         $user = User::findOrFail($id);
 
-        // Delete Previous Image
-        if (!is_null($user->image)) {
-            Storage::disk('public')->delete($user->image);
-        }
-
-        // Update New Info
+        // Update New Info without modifying the image
         $user->update($validated);
 
         return $user;
@@ -163,6 +176,11 @@ class UserController extends Controller
     public function password(UserRequest $request, string $id)
     {
         $user = User::findOrFail($id);
+        
+        // Validate old password
+        if (!Hash::check($request->old_password, $user->password)) {
+            return response()->json(['message' => 'Current password is incorrect.'], 422);
+        }
 
         $validated = $request->validated();
